+++ src/proto/mod.rs
@@ -1,5 +1,0 @@
-// @generated
-
-pub mod package;
-pub mod signed;
-pub mod versions;